+++ lib/bootstrap.scm
@@ -659,6 +659,24 @@
     (--> port (flush))))
 
 ;; -----------------------------------------------------------------------------
+(define (read-all . rest)
+  "(read-all)
+   (read-all port)
+
+   Read all S-Expressions from port and return then as a list"
+  (let ((port (if (null? rest) (current-input-port) (car rest))))
+    (do ((result '() (cons (read port) result)))
+      ((eof-object? (peek-char port)) (reverse result)))))
+
+;; -----------------------------------------------------------------------------
+(define (with-input-from-string string thunk)
+  "(with-input-from-string string thunk)
+
+   Reads string and execute the thunk with currect input port set to
+   port created from opening the string."
+  (with-input-from-port (open-input-string string) thunk))
+
+;; -----------------------------------------------------------------------------
 (define (regex? x)
   "(regex? x)