+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * add timestamp to error logs
 * make `+inf.0` and `-inf.0` parser constants
 * add `zip` procedure (inspired by [Lodash](https://docs-lodash.com/v4/zip/))
+* add `with-input-from-string` and `read-all` [#327](https://github.com/jcubic/lips/issues/327)
 ### Bugfix
 * fix `let-values` to allow binding to list [#281](https://github.com/jcubic/lips/issues/281)
 * fix wrong strings in `string-fill!`